Refactor Component ErrorMessages to tsx
- Message as children
- Make TS
- Move to components2
