+++ src/pages/Reports/pages/SubmitReport/components/LocateMeMap/index.js
@@ -13,7 +13,7 @@ import styled from 'styled-components';
 import FMBCredits from '~/apps/Map/components/FMBCredits';
 import LocatorControl from '~/apps/Map/components/LocatorControl';
 import AutocompleteGeocoder from '~/components/AutocompleteGeocoder';
-import ErrorMessage from '~/components/ErrorMessage';
+import { ErrorMessage } from '~/components2/ErrorMessage';
 import config from '~/pages/Reports/config';
 import { actions as errorStateActions } from '~/pages/Reports/state/ErrorState';
 import {
@@ -213,10 +213,9 @@ class LocateMeMap extends Component {
     return (
       <MapView>
         {this.props.error.message && (
-          <ErrorMessage
-            message={this.props.error.message}
-            onDismiss={this.props.removeError}
-          />
+          <ErrorMessage onDismiss={this.props.removeError}>
+            {this.props.error.message}
+          </ErrorMessage>
         )}
 
         {!this.state.isLoading && (
