+++ src/pages/Reports/pages/SubmitReport/pages/LocateModeChooser.js
@@ -6,7 +6,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import styled from 'styled-components';
 
-import ErrorMessage from '~/components/ErrorMessage';
+import { ErrorMessage } from '~/components2/ErrorMessage';
 import CloseIcon from '~/images/close.svg';
 import LocateIcon from '~/images/reports/location-button.svg';
 import MapIcon from '~/images/reports/noun-map-1909219.svg';
@@ -111,12 +111,13 @@ const LocateModeChooser = ({
     </CloseButton>
     {error.message && (
       <ErrorMessage
-        message={error.message}
         onDismiss={() => {
           removeError();
           onUseGeocoding();
         }}
-      />
+      >
+        {error.message}
+      </ErrorMessage>
     )}
   </Wrapper>
 );
